parser/tests/tree_construction/basic.dat

#data
<ul><li></li><div><li></div><li><li><div><li><address><li><b><em></b><li></ul>
#errors
(1,4): expected-doctype-but-got-start-tag
(1,45): end-tag-too-early
(1,58): end-tag-too-early
(1,69): adoption-agency-1.3
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|       <div>
|         <li>
|       <li>
|       <li>
|         <div>
|       <li>
|         <address>
|       <li>
|         <b>
|           <em>
|       <li>
